requirements.txt
pyswisseph>=2.10.3
pandas>=2.0.0
openpyxl>=3.1.2
matplotlib>=3.7.0
tqdm>=4.65.0
